+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## VERSION 4.48.0
+_11_06_2020_
+* FEATURE - experiments with installments row
+* ENHANCEMENT - persist payment configuration
+* FIX - top fragment position in default congrats order
+
+## VERSION 4.47.0
+_04_06_2020_
+* FEATURE - Send security type of payment as header of payment wrapper api call
+* ENHANCEMENT - Alternative order in congrats
+* ENHANCEMENT - Added amount to security validation data
+* ENHANCEMENT - Change account money logo
+* ENHANCEMENT - Take security required parameter from biometrics lib callback
+
 ## VERSION 4.46.4
 _12_06_2020_
 * ENHANCEMENT - Update CFTFormatter for CFTNA